# bpo-33234 Improve list() pre-sizing for inputs with known lengths