The list() constructor isn't taking full advantage of known input
lengths or length hints. This commit makes the constructor
pre-size and not over-allocate when the input size is known or
can be reasonably estimated.

A new test is added to `test_list.py` and a test needed to be modified
in `test_descr.py` as it assumed that there is only one call to
`__length_hint__()` in the list constructor.

<!-- issue-number: [bpo-33234](https://www.bugs.python.org/issue33234) -->
https://bugs.python.org/issue33234
<!-- /issue-number -->