@vstinner 2018-09-28
I'm not sure about this change because of the possible slowdown for short datasets. Raymond and me proposed to add a slot for length hint. I would suggest to work in two steps:

* Optimize list allocation only using the *length* if it's available (not the length hint)
* Later add a slot for length hint and use it for list allocation