@pablogsal 2018-09-28
Agreed. Will open a new PR with the version that only uses *lenght* if it is available.